src/main/java/com/seb/customer/http/validate/PhoneNumberValidator.java
package com.seb.customer.http.validate;

import com.google.i18n.phonenumbers.NumberParseException;
import com.google.i18n.phonenumbers.PhoneNumberUtil;
import com.google.i18n.phonenumbers.Phonenumber;

import javax.validation.ConstraintValidator;
import javax.validation.ConstraintValidatorContext;

import static java.util.regex.Pattern.compile;

public class PhoneNumberValidator implements ConstraintValidator<ValidPhoneNumber, String> {
    @Override
    public boolean isValid(String phoneNumber, ConstraintValidatorContext context) {
        try {
            PhoneNumberUtil util = PhoneNumberUtil.getInstance();
            Phonenumber.PhoneNumber parsedNumber = util.parse(phoneNumber, "");
            return util.isValidNumber(parsedNumber);
        } catch (NumberParseException e) {
            return false;
        }
    }

}
